Revert "#5749 fix tableDisplay memory leaks (#5886)"
This reverts commit 7c306ecda50afc6d66e5862dde1d4ea2314f42de.
