+++ beakerx/js/src/tableDisplay/tableModal/index.js
@@ -37,8 +37,6 @@ module.exports = function(TableScope) {
         self.hideModal();
       });
 
-    this.columnLimitModal = $modal;
-
     this.element.prepend($modal);
 
     if (this.columnNames.length > this.outputColumnLimit) {
